[Bug Fix] 1. 进一步完善了两个 daemon 进程的问题（子进程没有退出） 2. 修复了当前 user_id 不存在导致的 crawl_data attribute 不存在问题 3. 更新了使用了 Resource Manage 之后浏览器没有销毁的问题。

+++ CSDNVisualize/libCrawler/spider/daemonize_use_threadpool.py
@@ -395,12 +395,8 @@ if __name__ == '__main__':
         # Signal handler for termination (required)
         def sigterm_handler(signo, frame):
             ''' 如果要退出程序，清除 robot browsers'''
-            while True:
-                try:
-                    browser = browserQ.get(timeout=3)
-                    browser.quit()
-                except queue.Empty:
-                    break
+            res = Resource()
+            res.handler_quit()
 
             stop_and_free_thread_pool()
             # show_all_results()   # show_all_errors()
@@ -416,7 +412,8 @@ if __name__ == '__main__':
 
         # make sure parent-> chile processing exit
         ppid = os.getppid()
-        os.system("kill -15 {}".format(ppid))
+        if ppid != 1:
+            os.system("kill -15 {}".format(ppid))
 
         main(len(sys.argv), sys.argv)
